[core] fix implicit integer conversion warnings (#11709)
This commit enables the `-Wimplicit-int-conversion` compiler flag for
`ftd`, `mtd`, and `radio` to improve code quality by detecting
potential data loss from implicit type conversions. This is enabled
when clang toolchain is used.

All resulting warnings have been addressed by either:
- Changing variable, parameter, or return types to ensure consistency
  and prevent overflows.
- Adding explicit `static_cast` where the type conversion is intended
  and safe.

+++ src/core/border_router/routing_manager.hpp
@@ -893,7 +893,7 @@ private:
     class RdnssAddress
     {
     public:
-        void                SetFrom(const RecursiveDnsServerOption &aRdnss, uint8_t aAddressIndex);
+        void                SetFrom(const RecursiveDnsServerOption &aRdnss, uint16_t aAddressIndex);
         const Ip6::Address &GetAddress(void) const { return mAddress; }
         const TimeMilli    &GetLastUpdateTime(void) const { return mLastUpdateTime; }
         uint32_t            GetLifetime(void) const { return mLifetime; }
